add be

+++ src/Services/Estore.Domain/Models/Base/FileEntity.cs
@@ -6,7 +6,7 @@ namespace EStore.Domain.Models.Base;
 public class FileEntity : Entity<Guid>
 {
     public string FileName { get; set; } = default!;
-    public decimal FileSize { get; set; }
+    public long FileSize { get; set; }
     public FileType FileType { get; set; }
     public string Extension { get; set; } = default!;
     public string UserId { get; set; }
